Unrotated Dimensions on Mirabuf Scene Object
